+++ src/components/consumer/order/index.ts
@@ -1,0 +1,3 @@
+export { OrderCheckoutPanel } from './OrderCheckoutPanel';
+export { OrderProductSummary } from './OrderProductSummary';
+export { OrderProgressSteps } from './OrderProgressSteps';
